BSD/GNU portability fixes.
In GNU sed regular expressions '\|' is an extension even when not using
extended syntax, so use '[|]' to match a single pipe character. This
should work on all platforms as expected.

GNU mktemp needs at least 3 X's in the template.

md5 is not available on GNU/Linux out of the box, so use md5sum or md5,
whichever is present.

Also removed some unnecessary backslash escapes.